+++ diesel/src/types/impls/tuples.rs
@@ -2,6 +2,7 @@ use std::error::Error;
 
 use associations::BelongsTo;
 use backend::Backend;
+use deserialize;
 use expression::{AppearsOnTable, Expression, NonAggregate, SelectableExpression};
 use insertable::{CanInsertInSingleQuery, InsertValues, Insertable};
 use query_builder::*;
@@ -37,8 +38,6 @@ macro_rules! tuple_impls {
             impl<$($T),+, $($ST),+, DB> FromSqlRow<($($ST,)+), DB> for ($($T,)+) where
                 DB: Backend,
                 $($T: FromSqlRow<$ST, DB>),+,
-                $(DB: HasSqlType<$ST>),+,
-                DB: HasSqlType<($($ST,)+)>,
             {
                 const FIELDS_NEEDED: usize = $($T::FIELDS_NEEDED +)+ 0;
 
@@ -50,8 +49,6 @@ macro_rules! tuple_impls {
             impl<$($T),+, $($ST),+, DB> Queryable<($($ST,)+), DB> for ($($T,)+) where
                 DB: Backend,
                 $($T: Queryable<$ST, DB>),+,
-                $(DB: HasSqlType<$ST>),+,
-                DB: HasSqlType<($($ST,)+)>,
             {
                 type Row = ($($T::Row,)+);
 
@@ -65,7 +62,7 @@ macro_rules! tuple_impls {
                 DB: Backend,
                 $($T: QueryableByName<DB>,)+
             {
-                fn build<RowT: NamedRow<DB>>(row: &RowT) -> Result<Self, Box<Error + Send + Sync>> {
+                fn build<RowT: NamedRow<DB>>(row: &RowT) -> deserialize::Result<Self> {
                     Ok(($($T::build(row)?,)+))
                 }
             }